+++ src/PluginMetadata.cs
@@ -4,5 +4,5 @@ internal class PluginMetadata
 {
     public const string PLUGIN_GUID = "umno.QualityCompany";
     public const string PLUGIN_NAME = "QualityCompany";
-    public const string PLUGIN_VERSION = "1.2.2";
+    public const string PLUGIN_VERSION = "1.3.4";
 }