+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added ability to create a terminal for your mod using the name as an abbreviation (e.g. `QualityCompany` turns into `qc`), this can be customized within the `Register` call
 - Added ability to include your mod's terminal commands within the 'Help' command, can be disabled
 - Added more description options for sub commands to show when typing in the primary command
+- Fixed `hack` command to spawn items correctly and sync with clients
 
 # v1.3.2: Total Scrap UI Config