+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 - Added ability to include your mod's terminal commands within the 'Help' command, can be disabled
 - Added more description options for sub commands to show when typing in the primary command
 
+## Debug commands
+
+- Fixed `hack` command to spawn items correctly and sync with clients
+
 # v1.3.2: Total Scrap UI Config
 
 - Introduced config item to disable **Total Scrap UI** separate from item scrap UI
@@ -27,7 +31,7 @@
   - (worst case) Disable the feature in the config if it causes issues!
 - **[UNTESTED]** (hopefully) fully implemented ability to disable Networking operations in config
   - This will cause everything to be done CLIENT-SIDE ONLY
-  - This WILL cause desync issues with `sell` & `target` commands, *but* should recover on action of deposit desk eating the items
+  - This WILL cause desync issues with `sell` & `target` commands, _but_ should recover on action of deposit desk eating the items
   - Let us know if anything breaks from this please :)
 
 # v1.2.3: Mod Compatibility & Experimental fixes